chore: move form-contactPoint component
