+++ src/components/form-contactPoint/configured-form-contactPoint.jsx
@@ -2,7 +2,7 @@ import { reduxForm } from 'redux-form';
 
 import FormContactPoint from './form-contactPoint.component';
 import validate from './form-contactPoint-validations';
-import { asyncValidateDatasetInvokePatch } from '../formsLib/asyncValidateDatasetInvokePatch';
+import { asyncValidateDatasetInvokePatch } from '../../pages/dataset-registration-page/formsLib/asyncValidateDatasetInvokePatch';
 
 const config = {
   form: 'contactPoint',
