+++ src/components/form-contactPoint/form-contactPoint.component.tsx
@@ -5,12 +5,12 @@ import { Field } from 'redux-form';
 import {
   withTranslations,
   Props as TranslationsProps
-} from '../../../providers/translations';
+} from '../../providers/translations';
 
-import Helptext from '../../../components/helptext/helptext.component';
-import InputField from '../../../components/fields/field-input/field-input.component';
-import InputFieldReadonly from '../../../components/fields/field-input-readonly/field-input-readonly.component';
-import LinkReadonlyField from '../../../components/fields/field-link-readonly/field-link-readonly.component';
+import Helptext from '../helptext/helptext.component';
+import InputField from '../fields/field-input/field-input.component';
+import InputFieldReadonly from '../fields/field-input-readonly/field-input-readonly.component';
+import LinkReadonlyField from '../fields/field-link-readonly/field-link-readonly.component';
 
 interface ExternalProps {
   isReadOnly: boolean;
